Switch limit argument to Long
- Switched from Int to Long type
- Done for both validate and scan blocks functions
- Changed the default parameter for nullable one in the public interface to let callers be explicit about the argument
- Changed in the Rust layer as well

+++ sdk-lib/src/main/java/cash/z/ecc/android/sdk/block/CompactBlockProcessor.kt
@@ -882,7 +882,7 @@ class CompactBlockProcessor internal constructor(
         internal suspend fun validateBatchOfBlocks(batch: Batch, backend: Backend): BlockProcessingResult {
             Twig.verbose { "Starting to validate batch $batch" }
 
-            val result = backend.validateCombinedChainOrErrorBlockHeight(batch.range.length().toInt())
+            val result = backend.validateCombinedChainOrErrorBlockHeight(batch.range.length())
 
             return if (null == result) {
                 Twig.verbose { "Successfully validated batch $batch" }
@@ -895,7 +895,7 @@ class CompactBlockProcessor internal constructor(
         @VisibleForTesting
         @Suppress("MagicNumber")
         internal suspend fun scanBatchOfBlocks(batch: Batch, backend: Backend): BlockProcessingResult {
-            val scanResult = backend.scanBlocks(batch.range.length().toInt())
+            val scanResult = backend.scanBlocks(batch.range.length())
             return if (scanResult) {
                 Twig.verbose { "Successfully scanned batch $batch" }
                 BlockProcessingResult.Success
